refactor(archtest): 迁移 rmq 规则族 → module-path-agnostic Check* + 缩 baseline [#1636]
MOVE rmq_invariants_test.go 的 5 条 INVARIANT 检测逻辑（RMQ-CHANNEL-DESTRUCTION-VIA-CONN-01 /
RMQ-CHANNEL-MAX-PER-CONN-01 / RMQ-PUBLISHER-FAILURE-HANDLING-01 / RMQ-PUBLISHER-RELEASES-CHANNEL-01 /
RMQ-STOPINTAKE-INFLIGHT-WAIT-01）到非 test 的 rmq_invariants.go（Check* + scanner helpers），平台符号路径
经 PlatformModulePath 派生（line 77 bare github.com/ghbvf/gocell/adapters/rabbitmq → rmqAdapterPkgPath）；
_test.go 瘦身为 dogfood + 保留 _NamingImmunity / _ContainerCoverage_Wave4_RED 自检；baseline 删 1 行。

gocognit ≤15 经 checkPublishStmtViolations 的 7-arm switch 行为保持 helper 提取达成
（recursePublishStmtList / recursePublishCaseClauses，折叠 Block/For/Range 与 Switch/TypeSwitch 重复臂），
无 //nolint。register=no（gocell-internal-layout，扫 adapters/rabbitmq，外部 vacuous）。行为零变更
（verbatim MOVE + 恒等派生），TypedOpts 无 Tags 原样保留。

ref: golang.org/x/tools/go/analysis Analyzer
